+++ frontend/app/src/components/helper/display/icons/AssetIcon.vue
@@ -1,18 +1,34 @@
 <template>
-  <generated-icon
-    v-if="!!currency || error"
-    :asset="displayAsset"
-    :currency="!!currency"
-    :size="size"
-  />
-  <v-img
-    v-else-if="!error"
-    :src="url"
-    :max-width="size"
-    :min-width="size"
-    contain
-    @error="error = true"
-  />
+  <v-tooltip top open-delay="400">
+    <template #activator="{ on, attrs }">
+      <generated-icon
+        v-if="!!currency || error"
+        :asset="displayAsset"
+        :currency="!!currency"
+        :size="size"
+        v-bind="attrs"
+        v-on="on"
+      />
+      <v-img
+        v-else-if="!error"
+        :src="url"
+        :max-width="size"
+        :min-width="size"
+        contain
+        v-bind="attrs"
+        @error="error = true"
+        v-on="on"
+      />
+    </template>
+    <span>
+      {{
+        $t('asset_icon.tooltip', {
+          symbol: getSymbol(identifier),
+          name: getAssetName(identifier)
+        })
+      }}
+    </span>
+  </v-tooltip>
 </template>
 
 <script lang="ts">