+++ frontend/app/src/main.ts
@@ -1,5 +1,6 @@
 /* istanbul ignore file */
 
+import VueCompositionAPI, { provide } from '@vue/composition-api';
 import Vue from 'vue';
 import App from '@/App.vue';
 import '@/filters';
@@ -22,10 +23,14 @@ Vue.config.productionTip = false;
 
 Vue.use(Api);
 Vue.use(Interop);
+Vue.use(VueCompositionAPI);
 
 setupPremium();
 
 new Vue({
+  setup() {
+    provide('vuex-store', store);
+  },
   vuetify,
   router,
   store,